Adaptive Sparse Grids

Sparse grids can be adaptively constructed using a Gerstner–Griebel [4] type construction. This is implemented in the adaptive_sparsegrid function.

One dimensional example

To illustrate the adaptive algorithm consider functions that can be exactly approximated using a sparse grid approximation. Consider the polynomial $f:[-1,1]\mapsto\R$

\[f(x) = 3x^2 + 2x +1.\]

    using SparseGridsKit
    ndims = 1
    f(x) = @. 3.0*x[1]^2 + 2.0*x[1] + 1.0

f (generic function with 1 method)

The adaptive_sparsegrid function can be called with a function $f$ and the dimension of the function domain ndims.

domain = [[-1,1]]
(sg, f_on_Z) = adaptive_sparsegrid(f, domain, ndims)

(SparseGrid(1, [1 2], Any[[[1, 1]], [[2, 1]], [[2, 2]], [[2, 3]]], Integer[1; 2; 3;;], Integer[1, 2, 1, 3], Any[0, 1, 1, 1], Integer[0, 1], Any[Dict{Any, Any}([2, 3] => 3, [2, 1] => 2, [1, 1] => 1, [2, 2] => 1)], Any[[[0.0], [1.0, 0.0, -1.0]]], Any[[0.0, 1.0, -1.0]], [0.0; 1.0; -1.0;;], Any[Any[Any[Fun(Chebyshev(-1 .. 1), [1.0])], Any[Fun(Chebyshev(-1 .. 1), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1 .. 1), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1 .. 1), [0.25, -0.5, 0.25])]]], [[-1, 1]]), [1.0, 6.0, 2.0])

The function f can be exactly represented by a three point interpolant, which is identified in the adaptive algorithm.

test_points = range(-1,stop=1,length=100)
all([f(x)] ≈ interpolate_on_sparsegrid(sg,f_on_Z,x) for x in test_points),  get_n_grid_points(sg)

(true, 3)

Taking powers $k$ of the polynomial $f$ gives a polynomial $f^k$ of polynomial degree $2^k$. Consequently, this is represented exactly using $2^k+1$ interpolation points. The adaptive algorithm identifies this.

        f2(x) = f(x).^2
        (sg, f2_on_Z) = adaptive_sparsegrid(f2, domain, ndims)
        # Expect three point approx cubic (1 iteration to suffice)
        all([f2(x)] ≈ interpolate_on_sparsegrid(sg,f2_on_Z,x) for x in test_points),
        get_n_grid_points(sg)

(true, 5)

        f3(x) = f(x).^3
        (sg, f3_on_Z) = adaptive_sparsegrid(f3, domain, ndims)
        # Expect three point approx cubic (1 iteration to suffice)
        all([f3(x)] ≈ interpolate_on_sparsegrid(sg,f3_on_Z,x) for x in test_points),get_n_grid_points(sg)

(true, 9)

Multi-dimensional example

The adaptive algorithm is also applicable for higher dimensional functions. Consider the genz gaussianpeak example.

        using SparseGridsKit, LinearAlgebra
        n = 8
        C = 1.0
        W = 0.0
        T = "exponentialdecay"
        N = "gaussianpeak"
        f = genz(n::Int, C::Float64, W::Float64, T::String, N::String)
        domain = fill([-1,1],n)

8-element Vector{Vector{Int64}}:
 [-1, 1]
 [-1, 1]
 [-1, 1]
 [-1, 1]
 [-1, 1]
 [-1, 1]
 [-1, 1]
 [-1, 1]

An adaptive approximation is formed. This produces a (relatively) accurate approximation of the high-dimensional function $f$. The iterations halt when the stopping criterion of max profit p_max < proftol.

    # Approximate
    (sg, f_on_Z) = adaptive_sparsegrid(f, domain, n)

    nmc = 100
    test_points = [2 * rand(8) .- 1 for ii = 1:nmc]

    f_approx_on_test = interpolate_on_sparsegrid(sg,f_on_Z,test_points)

    norm([f(x) - f_approx_on_test[i] for (i,x) in enumerate(test_points)])

0.0012491954020144391

Reducing proftol allows further iterations and a more accurate approximation.

    (sg, f_on_Z) = adaptive_sparsegrid(f, domain, n, proftol=1e-5)

    f_approx_on_test = interpolate_on_sparsegrid(sg,f_on_Z,test_points)

    norm([f(x) - f_approx_on_test[i] for (i,x) in enumerate(test_points)])

0.00032486775966017

Functions

SparseGridsKit.adaptive_estimate — Method

adaptive_estimate(sg, sg_enhanced, f_on_z_enhanced, pcl)

Estimates the profit of adding multi-indices {α} in reduced margin to the sparse grid.

Arguments

sg: Sparse grid.

sg_enhanced: Enhanced sparse grid.

f_on_z_enhanced: Function evaluations on the enhanced sparse grid.

pcl: Precomputed Lagrange integrals.

rule: Level function(s) for sparse grid.

knots: Knot function(s) for sparse grid.

Returns

Vector of profits for each multi-index α.

SparseGridsKit.adaptive_mark — Function

adaptive_mark(α, p_α, θ=1e-4)

Dorfler style marking of multi-indices based on profits

Arguments

α: Multi-indices in reduced margin.

p_α: Vector of profits.

θ: Threshold for marking. Default is 1e-4.

SparseGridsKit.adaptive_refine — Method

adaptive_refine(sg, α_marked, rule, knots)

Refines the sparse grid based on marked multi-indices

Arguments

sg: Sparse grid.

α_marked: Marked multi-indices.

rule: Level function(s) for sparse grid.

knots: Knot function(s) for sparse grid.

Returns

sg: Refined sparse grid.

f_on_z: Function evaluations on the refined sparse grid.

SparseGridsKit.adaptive_solve — Method

adaptive_solve(f, sg,sg_enhanced, f_on_z)

Computes function f on the new grid points in sg_enhanced

Arguments

f: Function to be approximated.

sg: Sparse grid.

sg_enhanced: Enhanced sparse grid.

f_on_z: Function evaluations on the sparse grid sg.

Returns

Function evaluations on the enhanced sparse grid sg_enhanced.

SparseGridsKit.adaptive_sparsegrid — Method

adaptive_sparsegrid(f, ndims; maxpts = 100, proftol = 1e-4, rule=doubling, knots=ccpoints)

Constructs an adaptive sparse grid for approximating the function f in ndims dimensions.

Arguments

f: Function to be approximated taking arguments x with length(x)=ndims.

ndims: Dimension of domain.

maxpts: (Optional) Maximum number of points to include in the sparse grid. Default is 100.

proftol: (Optional) Tolerance for profits. Default is 1e-4.

rule: (Optional) Level function(s) for sparse grid. Default is doubling.

knots: (Optional) Knot function(s) for sparse grid. Default is ccpoints.

θ: (Optional) Threshold for marking. Default is 1e-4.

Returns

sg: Final sparse grid used to approximate f

f_on_z: Evaluations of f on grid points in sparse grid sg

SparseGridsKit.adaptive_terminate — Method

adaptiveterminate(sg, pα, maxpts, proftol)

Test profits and sparse grid to determine loop termination

Arguments

sg: Sparse grid.

p_α: Vector of profits.

maxpts: Maximum number of grid points allowed.

proftol: Profit tolerance.

Returns

Flag to indicate loop termination if maxpts has been reached, or proftol attained.

SparseGridsKit.compute_profit — Method

compute_profit(sg_α, f_diff_α, pcl)

Computes the "profit" of a sparse grid supplemented by a multi-index α

Arguments

sg_α: Enhanced sparse grid with α

f_diff_α: Function differences (surpluses) at the enhanced sparse grid points.

pcl: Precomputed Lagrange integrals

Returns

Computed profit as Expected change in approximation.