Multi Index Sets

Sparse grids are constructed via multi-index sets specifying a cleverly chosen linear combination of tensor product grids.

Suppose we consider a function $u$ with domain $\Gamma\in\mathbb{R}^n$.

Tensor Product Multi-Index Sets

The multi-index set

\[\{\alpha : \Vert \alpha \Vert_\infty \leq n + d\} \subset \N^{d}\]

is constructed using the create_tensor_miset function.

using SparseGridsKit

n, k = 2, 1
mi_set = create_tensor_miset(n, k)

MISet([[1, 1], [1, 2], [2, 1], [2, 2]])

The multi-index set can be viewed a vector of Vector{Integer} representing each multi-index:

get_mi(mi_set)

4-element Vector{Vector{Int64}}:
 [1, 1]
 [1, 2]
 [2, 1]
 [2, 2]

Smolyak Multi-Index Sets

The standard Smolyak construction are also easily constructed. The multi-index set

\[\{\alpha : \Vert \alpha \Vert_1 \leq n + d\} \subset \N^{2}\]

is constructed using the following syntax.

using SparseGridsKit

n, k = 2, 1
mi_set = create_smolyak_miset(n, k)

MISet([[1, 1], [1, 2], [2, 1]])

This gives a different, smaller multi-index set than the tensor product multi-index set.

get_mi(mi_set)

3-element Vector{Vector{Int64}}:
 [1, 1]
 [1, 2]
 [2, 1]

Manipulating Multi-Index Sets

The multi-index set can be altered by adding additional multi-indices. Continuing using the above Smolyak example, we add the multi-index $[1, 3]$ with the add_mi function.

mi_set_new = MISet([[1,3]])
combined_mi_set = add_mi(mi_set, mi_set_new)

MISet([[1, 1], [1, 2], [1, 3], [2, 1]])

In Gerstner–Griebel style adaptive algorithms, the margin and reduced margin are generally required. These are easily computed using the get_margin and get_reduced_margin functions.

margin_set = get_margin(combined_mi_set)

MISet([[1, 4], [2, 2], [2, 3], [3, 1]])

reduced_margin_set = get_reduced_margin(combined_mi_set)

MISet([[1, 4], [2, 2], [3, 1]])

Notice how the multi-index $[2,3]$ is in the margin but not the reduced margin. Its backwards neighbour $[2,2]$ is missing from combined_mi_set.

Function Reference

SparseGridsKit.add_mi — Method

add_mi(mi_set, mi_set_new)

Adds a new set of multi-indices (mi_set_new) to an existing multi-index set (mi_set).

Arguments

mi_set: The original multi-index set.

mi_set_new: A new MISet to be added to the original.

Returns

A new MISet containing the combined and sorted multi-indices.

SparseGridsKit.add_mi — Method

add_mi(mi_set, mi_new)

Adds a single new multi-index (mi_new) to an existing multi-index set (mi_set).

Arguments

mi_set: The original multi-index set.

mi_new: A new multi-index vector to be added.

Returns

A new MISet containing the updated and sorted multi-indices.

SparseGridsKit.create_box_miset — Method

create_box_miset(n, k)

Creates a rectangular multi-index set for a given dimension (n) and levels vector (k).

Arguments

n: The dimensionality of the space.

k: A vector of levels for each dimension.

SparseGridsKit.create_rule_miset — Method

create_rule_miset(n, k, rule)

Creates a multi-index set for a given dimension (n) and with rule(mi) less than or equal to k.

Arguments

n: The dimensionality of the space.

k: Maximum for rule(mi)

rule: Rule to compute on each MI

SparseGridsKit.create_smolyak_miset — Method

create_smolyak_miset(n, k)

Creates a Smolyak multi-index set for a given dimension (n) and level (k).

Arguments

n: The dimensionality of the space.

k: The level of the Smolyak grid.

Returns

An MISet representing the Smolyak multi-index set.

SparseGridsKit.create_tensor_miset — Method

create_tensor_miset(n, k)

Creates a tensor product multi-index set for a given dimension (n) and level (k).

Arguments

n: The dimensionality of the space.

k: The level of the tensor product grid.

Returns

An MISet representing the tensor product multi-index set.

SparseGridsKit.create_totaldegree_miset — Method

create_totaldegree_miset(n, k)

Creates a multi-index set for a given dimension (n) and with total levels less than or equal to k.

Arguments

n: The dimensionality of the space.

k: Total degree

SparseGridsKit.get_margin — Method

get_margin(mi_set)

Calculates the margin of a multi-index set (mi_set), which is the set of multi-indices that can extend the current set.

Arguments

mi_set: An instance of MISet.

Returns

An MISet containing the margin set.

SparseGridsKit.get_mi — Method

get_mi(mi_set)

Retrieves the list of multi-indices from a given MISet.

Arguments

mi_set: An instance of MISet.

Returns

A vector of multi-index vectors.

SparseGridsKit.get_n_mi — Method

get_n_mi(mi_set)

Returns the number of multi-indices in a given MISet.

Arguments

mi_set: An instance of MISet.

Returns

The number of multi-indices in the set.

SparseGridsKit.get_reduced_margin — Method

get_reduced_margin(mi_set)

Calculates the reduced margin of a multi-index set (mi_set), which removes indices already in the set.

Arguments

mi_set: An instance of MISet.

Returns

An MISet containing the reduced margin set.