Spectral Sparse Grids

The sparse grid approximations are polynomial approximations, with a basis of Lagrange interpolation polynomials. The approximation can be written in any suitable polynomial basis, for example we could use the Chebyshev polynomials.

The change of basis is implemented exploiting the ApproxFun.jl package. The underlying interpolation polynomials used in SparseGridsKit.jl are constructed using Funs. To change basis we use an ApproxFun.jl expansion into coefficients and form a tensor product. To promote sparsity this is done using a truncated Kronecker product. The polynomial coefficients are assembled with corresponding polynomial degrees and spaces in a callable SpectralSparseGridApproximation.

Examples

An adaptive sparse grid approximation for a Genz test function is created.

using SparseGridsKit
n = 8
C = 1.0
W = 0.0
T = "exponentialdecay"
N = "gaussianpeak"
f = genz(n::Int, C::Float64, W::Float64, T::String, N::String)
domain = fill([-1,1],n)
# Approximate
(sg, f_on_Z) = adaptive_sparsegrid(f, domain, n)
f_sg = SparseGridApproximation(sg,f_on_Z)

SparseGridApproximation(SparseGrid(8, [1 1 … 3 4; 1 1 … 2 1; … ; 1 1 … 1 1; 1 1 … 1 1], Any[[[1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[1, 1], [1, 1], [2, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[1, 1], [1, 1], [2, 2], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[1, 1], [1, 1], [2, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[3, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[3, 2], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[3, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[3, 4], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[3, 5], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[3, 1], [2, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]]  …  [[3, 5], [2, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 2], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 3], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 4], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 5], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 6], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 7], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 8], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]], [[4, 9], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1], [1, 1]]], Integer[1 1 … 1 1; 1 1 … 1 1; … ; 8 1 … 1 1; 9 1 … 1 1], Integer[1, 2, 1, 3, 4, 5, 1, 6, 7, 8  …  17, 4, 18, 5, 19, 1, 20, 6, 21, 7], Any[-1, 1, 1, 1, -1, -1, -1, -1, -1, 1  …  1, 1, 1, 1, 1, 1, 1, 1, 1, 1], Integer[-1, 1, 0, 0, 0, -1, 1, 1], Any[Dict{Any, Any}([4, 2] => 6, [1, 1] => 1, [3, 5] => 3, [4, 5] => 1, [2, 1] => 2, [3, 1] => 2, [4, 8] => 9, [4, 1] => 2, [4, 6] => 8, [2, 3] => 3…), Dict{Any, Any}([4, 2] => 6, [1, 1] => 1, [3, 5] => 3, [4, 5] => 1, [2, 1] => 2, [3, 1] => 2, [4, 8] => 9, [4, 1] => 2, [4, 6] => 8, [2, 3] => 3…), Dict{Any, Any}([4, 2] => 6, [1, 1] => 1, [3, 5] => 3, [4, 5] => 1, [2, 1] => 2, [3, 1] => 2, [4, 8] => 9, [4, 1] => 2, [4, 6] => 8, [2, 3] => 3…), Dict{Any, Any}([4, 2] => 6, [1, 1] => 1, [3, 5] => 3, [4, 5] => 1, [2, 1] => 2, [3, 1] => 2, [4, 8] => 9, [4, 1] => 2, [4, 6] => 8, [2, 3] => 3…), Dict{Any, Any}([4, 2] => 6, [1, 1] => 1, [3, 5] => 3, [4, 5] => 1, [2, 1] => 2, [3, 1] => 2, [4, 8] => 9, [4, 1] => 2, [4, 6] => 8, [2, 3] => 3…), Dict{Any, Any}([4, 2] => 6, [1, 1] => 1, [3, 5] => 3, [4, 5] => 1, [2, 1] => 2, [3, 1] => 2, [4, 8] => 9, [4, 1] => 2, [4, 6] => 8, [2, 3] => 3…), Dict{Any, Any}([4, 2] => 6, [1, 1] => 1, [3, 5] => 3, [4, 5] => 1, [2, 1] => 2, [3, 1] => 2, [4, 8] => 9, [4, 1] => 2, [4, 6] => 8, [2, 3] => 3…), Dict{Any, Any}([4, 2] => 6, [1, 1] => 1, [3, 5] => 3, [4, 5] => 1, [2, 1] => 2, [3, 1] => 2, [4, 8] => 9, [4, 1] => 2, [4, 6] => 8, [2, 3] => 3…)], Any[[[0.0], [1.0, 0.0, -1.0], [1.0, 0.7071067811865475, 0.0, -0.7071067811865475, -1.0], [1.0, 0.9238795325112867, 0.7071067811865475, 0.3826834323650898, 0.0, -0.3826834323650898, -0.7071067811865475, -0.9238795325112867, -1.0]], [[0.0], [1.0, 0.0, -1.0], [1.0, 0.7071067811865475, 0.0, -0.7071067811865475, -1.0], [1.0, 0.9238795325112867, 0.7071067811865475, 0.3826834323650898, 0.0, -0.3826834323650898, -0.7071067811865475, -0.9238795325112867, -1.0]], [[0.0], [1.0, 0.0, -1.0], [1.0, 0.7071067811865475, 0.0, -0.7071067811865475, -1.0], [1.0, 0.9238795325112867, 0.7071067811865475, 0.3826834323650898, 0.0, -0.3826834323650898, -0.7071067811865475, -0.9238795325112867, -1.0]], [[0.0], [1.0, 0.0, -1.0], [1.0, 0.7071067811865475, 0.0, -0.7071067811865475, -1.0], [1.0, 0.9238795325112867, 0.7071067811865475, 0.3826834323650898, 0.0, -0.3826834323650898, -0.7071067811865475, -0.9238795325112867, -1.0]], [[0.0], [1.0, 0.0, -1.0], [1.0, 0.7071067811865475, 0.0, -0.7071067811865475, -1.0], [1.0, 0.9238795325112867, 0.7071067811865475, 0.3826834323650898, 0.0, -0.3826834323650898, -0.7071067811865475, -0.9238795325112867, -1.0]], [[0.0], [1.0, 0.0, -1.0], [1.0, 0.7071067811865475, 0.0, -0.7071067811865475, -1.0], [1.0, 0.9238795325112867, 0.7071067811865475, 0.3826834323650898, 0.0, -0.3826834323650898, -0.7071067811865475, -0.9238795325112867, -1.0]], [[0.0], [1.0, 0.0, -1.0], [1.0, 0.7071067811865475, 0.0, -0.7071067811865475, -1.0], [1.0, 0.9238795325112867, 0.7071067811865475, 0.3826834323650898, 0.0, -0.3826834323650898, -0.7071067811865475, -0.9238795325112867, -1.0]], [[0.0], [1.0, 0.0, -1.0], [1.0, 0.7071067811865475, 0.0, -0.7071067811865475, -1.0], [1.0, 0.9238795325112867, 0.7071067811865475, 0.3826834323650898, 0.0, -0.3826834323650898, -0.7071067811865475, -0.9238795325112867, -1.0]]], Any[[0.0, 1.0, -1.0, 0.7071067811865475, -0.7071067811865475, 0.9238795325112867, 0.3826834323650898, -0.3826834323650898, -0.9238795325112867], [0.0, 1.0, -1.0, 0.7071067811865475, -0.7071067811865475, 0.9238795325112867, 0.3826834323650898, -0.3826834323650898, -0.9238795325112867], [0.0, 1.0, -1.0, 0.7071067811865475, -0.7071067811865475, 0.9238795325112867, 0.3826834323650898, -0.3826834323650898, -0.9238795325112867], [0.0, 1.0, -1.0, 0.7071067811865475, -0.7071067811865475, 0.9238795325112867, 0.3826834323650898, -0.3826834323650898, -0.9238795325112867], [0.0, 1.0, -1.0, 0.7071067811865475, -0.7071067811865475, 0.9238795325112867, 0.3826834323650898, -0.3826834323650898, -0.9238795325112867], [0.0, 1.0, -1.0, 0.7071067811865475, -0.7071067811865475, 0.9238795325112867, 0.3826834323650898, -0.3826834323650898, -0.9238795325112867], [0.0, 1.0, -1.0, 0.7071067811865475, -0.7071067811865475, 0.9238795325112867, 0.3826834323650898, -0.3826834323650898, -0.9238795325112867], [0.0, 1.0, -1.0, 0.7071067811865475, -0.7071067811865475, 0.9238795325112867, 0.3826834323650898, -0.3826834323650898, -0.9238795325112867]], [0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0; … ; -0.3826834323650898 0.0 … 0.0 0.0; -0.9238795325112867 0.0 … 0.0 0.0], Any[Any[Any[Fun(Chebyshev(-1.0 .. 1.0), [1.0])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.5, 0.25])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.25, 0.25, 0.25, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.353553, -0.0, -0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.0, -0.5, 0.0, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.353553, -0.0, 0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.25, 0.25, -0.25, 0.125])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.0625, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.0625]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.23097, 0.176777, 0.0956709, -6.79875e-17, -0.0956709, -0.176777, -0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.176777, 4.61912e-17, -0.176777, -0.25, -0.176777, -4.61912e-17, 0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.0956709, -0.176777, -0.23097, 7.86817e-17, 0.23097, 0.176777, -0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 9.20989e-18, -0.25, -3.92523e-17, 0.25, 3.00424e-17, -0.25, 0.0, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.0956709, -0.176777, 0.23097, 6.58603e-17, -0.23097, 0.176777, 0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.176777, -6.93889e-18, 0.176777, -0.25, 0.176777, 6.93889e-18, -0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.23097, 0.176777, -0.0956709, -6.26767e-17, 0.0956709, -0.176777, 0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.0625, -0.125, 0.125, -0.125, 0.125, -0.125, 0.125, -0.125, 0.0625])]], Any[Any[Fun(Chebyshev(-1.0 .. 1.0), [1.0])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.5, 0.25])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.25, 0.25, 0.25, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.353553, -0.0, -0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.0, -0.5, 0.0, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.353553, -0.0, 0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.25, 0.25, -0.25, 0.125])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.0625, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.0625]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.23097, 0.176777, 0.0956709, -6.79875e-17, -0.0956709, -0.176777, -0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.176777, 4.61912e-17, -0.176777, -0.25, -0.176777, -4.61912e-17, 0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.0956709, -0.176777, -0.23097, 7.86817e-17, 0.23097, 0.176777, -0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 9.20989e-18, -0.25, -3.92523e-17, 0.25, 3.00424e-17, -0.25, 0.0, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.0956709, -0.176777, 0.23097, 6.58603e-17, -0.23097, 0.176777, 0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.176777, -6.93889e-18, 0.176777, -0.25, 0.176777, 6.93889e-18, -0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.23097, 0.176777, -0.0956709, -6.26767e-17, 0.0956709, -0.176777, 0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.0625, -0.125, 0.125, -0.125, 0.125, -0.125, 0.125, -0.125, 0.0625])]], Any[Any[Fun(Chebyshev(-1.0 .. 1.0), [1.0])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.5, 0.25])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.25, 0.25, 0.25, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.353553, -0.0, -0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.0, -0.5, 0.0, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.353553, -0.0, 0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.25, 0.25, -0.25, 0.125])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.0625, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.0625]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.23097, 0.176777, 0.0956709, -6.79875e-17, -0.0956709, -0.176777, -0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.176777, 4.61912e-17, -0.176777, -0.25, -0.176777, -4.61912e-17, 0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.0956709, -0.176777, -0.23097, 7.86817e-17, 0.23097, 0.176777, -0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 9.20989e-18, -0.25, -3.92523e-17, 0.25, 3.00424e-17, -0.25, 0.0, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.0956709, -0.176777, 0.23097, 6.58603e-17, -0.23097, 0.176777, 0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.176777, -6.93889e-18, 0.176777, -0.25, 0.176777, 6.93889e-18, -0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.23097, 0.176777, -0.0956709, -6.26767e-17, 0.0956709, -0.176777, 0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.0625, -0.125, 0.125, -0.125, 0.125, -0.125, 0.125, -0.125, 0.0625])]], Any[Any[Fun(Chebyshev(-1.0 .. 1.0), [1.0])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.5, 0.25])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.25, 0.25, 0.25, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.353553, -0.0, -0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.0, -0.5, 0.0, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.353553, -0.0, 0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.25, 0.25, -0.25, 0.125])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.0625, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.0625]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.23097, 0.176777, 0.0956709, -6.79875e-17, -0.0956709, -0.176777, -0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.176777, 4.61912e-17, -0.176777, -0.25, -0.176777, -4.61912e-17, 0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.0956709, -0.176777, -0.23097, 7.86817e-17, 0.23097, 0.176777, -0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 9.20989e-18, -0.25, -3.92523e-17, 0.25, 3.00424e-17, -0.25, 0.0, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.0956709, -0.176777, 0.23097, 6.58603e-17, -0.23097, 0.176777, 0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.176777, -6.93889e-18, 0.176777, -0.25, 0.176777, 6.93889e-18, -0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.23097, 0.176777, -0.0956709, -6.26767e-17, 0.0956709, -0.176777, 0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.0625, -0.125, 0.125, -0.125, 0.125, -0.125, 0.125, -0.125, 0.0625])]], Any[Any[Fun(Chebyshev(-1.0 .. 1.0), [1.0])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.5, 0.25])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.25, 0.25, 0.25, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.353553, -0.0, -0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.0, -0.5, 0.0, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.353553, -0.0, 0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.25, 0.25, -0.25, 0.125])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.0625, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.0625]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.23097, 0.176777, 0.0956709, -6.79875e-17, -0.0956709, -0.176777, -0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.176777, 4.61912e-17, -0.176777, -0.25, -0.176777, -4.61912e-17, 0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.0956709, -0.176777, -0.23097, 7.86817e-17, 0.23097, 0.176777, -0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 9.20989e-18, -0.25, -3.92523e-17, 0.25, 3.00424e-17, -0.25, 0.0, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.0956709, -0.176777, 0.23097, 6.58603e-17, -0.23097, 0.176777, 0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.176777, -6.93889e-18, 0.176777, -0.25, 0.176777, 6.93889e-18, -0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.23097, 0.176777, -0.0956709, -6.26767e-17, 0.0956709, -0.176777, 0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.0625, -0.125, 0.125, -0.125, 0.125, -0.125, 0.125, -0.125, 0.0625])]], Any[Any[Fun(Chebyshev(-1.0 .. 1.0), [1.0])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.5, 0.25])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.25, 0.25, 0.25, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.353553, -0.0, -0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.0, -0.5, 0.0, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.353553, -0.0, 0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.25, 0.25, -0.25, 0.125])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.0625, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.0625]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.23097, 0.176777, 0.0956709, -6.79875e-17, -0.0956709, -0.176777, -0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.176777, 4.61912e-17, -0.176777, -0.25, -0.176777, -4.61912e-17, 0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.0956709, -0.176777, -0.23097, 7.86817e-17, 0.23097, 0.176777, -0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 9.20989e-18, -0.25, -3.92523e-17, 0.25, 3.00424e-17, -0.25, 0.0, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.0956709, -0.176777, 0.23097, 6.58603e-17, -0.23097, 0.176777, 0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.176777, -6.93889e-18, 0.176777, -0.25, 0.176777, 6.93889e-18, -0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.23097, 0.176777, -0.0956709, -6.26767e-17, 0.0956709, -0.176777, 0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.0625, -0.125, 0.125, -0.125, 0.125, -0.125, 0.125, -0.125, 0.0625])]], Any[Any[Fun(Chebyshev(-1.0 .. 1.0), [1.0])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.5, 0.25])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.25, 0.25, 0.25, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.353553, -0.0, -0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.0, -0.5, 0.0, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.353553, -0.0, 0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.25, 0.25, -0.25, 0.125])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.0625, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.0625]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.23097, 0.176777, 0.0956709, -6.79875e-17, -0.0956709, -0.176777, -0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.176777, 4.61912e-17, -0.176777, -0.25, -0.176777, -4.61912e-17, 0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.0956709, -0.176777, -0.23097, 7.86817e-17, 0.23097, 0.176777, -0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 9.20989e-18, -0.25, -3.92523e-17, 0.25, 3.00424e-17, -0.25, 0.0, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.0956709, -0.176777, 0.23097, 6.58603e-17, -0.23097, 0.176777, 0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.176777, -6.93889e-18, 0.176777, -0.25, 0.176777, 6.93889e-18, -0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.23097, 0.176777, -0.0956709, -6.26767e-17, 0.0956709, -0.176777, 0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.0625, -0.125, 0.125, -0.125, 0.125, -0.125, 0.125, -0.125, 0.0625])]], Any[Any[Fun(Chebyshev(-1.0 .. 1.0), [1.0])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.5, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.5, -0.0, -0.5]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.5, 0.25])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.25, 0.25, 0.25, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, 0.353553, -0.0, -0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.0, -0.5, 0.0, 0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.25, -0.353553, -0.0, 0.353553, -0.25]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.25, 0.25, -0.25, 0.125])], Any[Fun(Chebyshev(-1.0 .. 1.0), [0.0625, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.125, 0.0625]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.23097, 0.176777, 0.0956709, -6.79875e-17, -0.0956709, -0.176777, -0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.176777, 4.61912e-17, -0.176777, -0.25, -0.176777, -4.61912e-17, 0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 0.0956709, -0.176777, -0.23097, 7.86817e-17, 0.23097, 0.176777, -0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, 9.20989e-18, -0.25, -3.92523e-17, 0.25, 3.00424e-17, -0.25, 0.0, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.0956709, -0.176777, 0.23097, 6.58603e-17, -0.23097, 0.176777, 0.0956709, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.176777, -6.93889e-18, 0.176777, -0.25, 0.176777, 6.93889e-18, -0.176777, 0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.125, -0.23097, 0.176777, -0.0956709, -6.26767e-17, 0.0956709, -0.176777, 0.23097, -0.125]), Fun(Chebyshev(-1.0 .. 1.0), [0.0625, -0.125, 0.125, -0.125, 0.125, -0.125, 0.125, -0.125, 0.0625])]]], Any[[-1.0, 1.0], [-1.0, 1.0], [-1.0, 1.0], [-1.0, 1.0], [-1.0, 1.0], [-1.0, 1.0], [-1.0, 1.0], [-1.0, 1.0]]), Any[1.0, 0.9986316107749044, 0.9986316107749044, 0.5420585004051459, 0.7362462226763177, 0.7362462226763177, 0.5420585004051459, 0.5265868123957612, 0.715231937600404, 0.9714575308793776  …  0.5265868123957612, 0.5265868123957612, 0.715231937600404, 0.9714575308793776, 0.715231937600404, 0.5265868123957612, 0.5929174027357444, 0.9142226183682018, 0.9142226183682018, 0.5929174027357444])

The approximation f_sg is a representation formed as a linear combination of Lagrange interpolation polynomials. This is converted to a spectral representation using convert_to_spectral_approximation.

f_spectral = convert_to_spectral_approximation(sg, f_on_Z)

SpectralSparseGridApproximation(8, [9, 3, 3, 1, 1, 1, 1, 1], Any[Chebyshev(-1.0 .. 1.0), Chebyshev(-1.0 .. 1.0), Chebyshev(-1.0 .. 1.0), Chebyshev(-1.0 .. 1.0), Chebyshev(-1.0 .. 1.0), Chebyshev(-1.0 .. 1.0), Chebyshev(-1.0 .. 1.0), Chebyshev(-1.0 .. 1.0)], sparsevec([1, 3, 4, 7, 13, 16, 19, 22, 25, 28, 37, 43, 46, 55, 73], Any[0.742164337972932, -0.0006841946125477893, -3.469446951953614e-17, -0.010755340910229266, 1.3877787807814457e-17, 6.938893903907228e-18, -0.2248162280230623, 1.3877787807814457e-17, 0.00326769527780945, -1.3877787807814457e-17, 0.017143315390855104, -0.0002481983722724887, -2.7755575615628914e-17, -0.0008868264965552902, 3.3862943736077955e-5], 81), sparse([1, 3, 4, 7, 13, 16, 19, 22, 25, 28, 37, 43, 46, 55, 73], [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1], [[0, 0, 0, 0, 0, 0, 0, 0], [0, 0, 2, 0, 0, 0, 0, 0], [0, 1, 0, 0, 0, 0, 0, 0], [0, 2, 0, 0, 0, 0, 0, 0], [1, 1, 0, 0, 0, 0, 0, 0], [1, 2, 0, 0, 0, 0, 0, 0], [2, 0, 0, 0, 0, 0, 0, 0], [2, 1, 0, 0, 0, 0, 0, 0], [2, 2, 0, 0, 0, 0, 0, 0], [3, 0, 0, 0, 0, 0, 0, 0], [4, 0, 0, 0, 0, 0, 0, 0], [4, 2, 0, 0, 0, 0, 0, 0], [5, 0, 0, 0, 0, 0, 0, 0], [6, 0, 0, 0, 0, 0, 0, 0], [8, 0, 0, 0, 0, 0, 0, 0]], 81, 1))

This still has an n dimensional domain.

f_spectral.dims

8

The representation of the tensor of polynomials is compactly stored as a SparseVector. To index into this we require the storage dimension for each domain dimension.

f_spectral.expansiondimensions

8-element Vector{Int64}:
 9
 3
 3
 1
 1
 1
 1
 1

This demonstrates that more polynomial terms are used in the parameter dimensions closer to one. Intuitively, this matches the anisotropic nature of the test function (see genz). The polynomial coefficients and the corresponding polynomial degrees are stored in sparse vectors.

f_spectral.coefficients, f_spectral.polydegrees

(sparsevec([1, 3, 4, 7, 13, 16, 19, 22, 25, 28, 37, 43, 46, 55, 73], Any[0.742164337972932, -0.0006841946125477893, -3.469446951953614e-17, -0.010755340910229266, 1.3877787807814457e-17, 6.938893903907228e-18, -0.2248162280230623, 1.3877787807814457e-17, 0.00326769527780945, -1.3877787807814457e-17, 0.017143315390855104, -0.0002481983722724887, -2.7755575615628914e-17, -0.0008868264965552902, 3.3862943736077955e-5], 81), sparse([1, 3, 4, 7, 13, 16, 19, 22, 25, 28, 37, 43, 46, 55, 73], [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1], [[0, 0, 0, 0, 0, 0, 0, 0], [0, 0, 2, 0, 0, 0, 0, 0], [0, 1, 0, 0, 0, 0, 0, 0], [0, 2, 0, 0, 0, 0, 0, 0], [1, 1, 0, 0, 0, 0, 0, 0], [1, 2, 0, 0, 0, 0, 0, 0], [2, 0, 0, 0, 0, 0, 0, 0], [2, 1, 0, 0, 0, 0, 0, 0], [2, 2, 0, 0, 0, 0, 0, 0], [3, 0, 0, 0, 0, 0, 0, 0], [4, 0, 0, 0, 0, 0, 0, 0], [4, 2, 0, 0, 0, 0, 0, 0], [5, 0, 0, 0, 0, 0, 0, 0], [6, 0, 0, 0, 0, 0, 0, 0], [8, 0, 0, 0, 0, 0, 0, 0]], 81, 1))

The spectral representation is the same function as the sparse grid representation.

x_test = [rand(n) for i in 1:100]
y_test = f_sg.(x_test)
y_spectral = f_spectral.(x_test)
all(isapprox(y_test, y_spectral; atol=1e-8))

true

The spectral sparse grid approximation also supports addition and subtraction.

f_spectral_2 = f_spectral + f_spectral
y_spectral_2 = f_spectral_2.(x_test)

all(isapprox(y_spectral_2, 2*y_spectral; atol=1e-8))

true

f_spectral_0 = f_spectral - f_spectral
y_spectral_0 = f_spectral_0.(x_test)

all(isapprox(y_spectral_0, zeros(size(y_spectral_0)); atol=1e-8))

false

Function Reference

SparseGridsKit.SpectralSparseGridApproximation — Type

SpectralSparseGridApproximation

A spectral sparse grid approximation.

Fields

dims::Int: Dimension of the domain.

expansiondimensions::Vector{Int}: Vector of maximum polynomial degree in each dimension.

polytypes::Vector{Any}: Vector of polynomial spaces.

coefficients::SparseVector{Any}: Sparse ector of polynomial coefficients

polydegrees::SparseVector{Vector}: SparseVector of polynomial degrees for each non zero coefficient.

SparseGridsKit.SpectralSparseGridApproximation — Method

SpectralSparseGridApproximation(dims, expansiondims, polytypes, coefficients)

Constructs a spectral sparse grid approximation.

Arguments

dims::Int: Dimension of the domain.

expansiondimensions::Vector{Int}: Vector of maximum polynomial degree in each dimension.

polytypes::Vector{Any}: Vector of polynomial spaces.

coefficients::Vector{Any}: Kronecker product vector of polynomial coefficients

Returns

SpectralSparseGridApproximation: An instance of SpectralSparseGridApproximation.

SparseGridsKit.SpectralSparseGridApproximation — Method

(s::SpectralSparseGridApproximation)(x)

Evaluate the spectral sparse grid approximation at the given point x.

Arguments

s::SpectralSparseGridApproximation: Spectral sparse grid approximation object.

x::Vector{T}: Vector x in domain to be evaluated

Returns

evaluation: Evaluation of s(x)

Base.:+ — Method

+(grid1::SpectralSparseGridApproximation, grid2::SpectralSparseGridApproximation)

Add two SpectralSparseGridApproximation objects.

Arguments

grid1::SpectralSparseGridApproximation: Spectral sparse grid approximation.

grid2::SpectralSparseGridApproximation: Second spectral sparse grid approximation.

Returns

SpectralSparseGridApproximation representing sum of grid1 and grid2.

Base.:- — Method

-(grid1::SpectralSparseGridApproximation, grid2::SpectralSparseGridApproximation)

Subtracts SpectralSparseGridApproximation objects

Arguments

grid1::SpectralSparseGridApproximation: Spectral sparse grid approximation.

grid2::SpectralSparseGridApproximation: Second spectral sparse grid approximation, to be subtracted

Returns

SpectralSparseGridApproximation object representing the subtraction.

SparseGridsKit.convert_to_spectral_approximation — Method

convert_to_spectral_approximation(sparsegrid::SparseGrid, fongrid)

Converts a sparse grid and corresponding evaluations fongrid to a SpectralSparseGridApproximation

Arguments

sparsegrid::SparseGrid: Sparse grid

fongrid: Function evaluations of sparse grid.

Returns

SpectralSparseGridApproximation represntation of sparsegrid with evaluations fongrid.

SparseGridsKit.convert_to_spectral_approximation — Method

convert_to_spectral_approximation(sga::SparseGridApproximation)

Convert a SparseGridApproximation object to its spectral approximation.

Arguments

sga::SparseGridApproximation: Sparse grid approximation.

Returns

Spectral Sparse Grid Approximation representing sga.

SparseGridsKit.createlagrangepolys_approxfun — Function

createlagrangepolys_approxfun(pts)

Create a vector of Lagrange polynomial approximation functions for a given set of knots.

Arguments

pts: Vector of knots.

Returns

Vector{Any}: Vector containing the Lagrange polynomial approximation Funs for each point in pts.

SparseGridsKit.kron_index — Method

kron_index(indices::Vector{Int}, dims::Vector{Int})

Calculate the linear index in a Kronecker product space given the multi-dimensional indices and the dimensions of each space.

Arguments

indices::Vector{Int}: Index vector.

dims::Vector{Int}: Vector of storage dimension for each tensor dimension.

Returns

Linear index corresponding to the index given in indices.

SparseGridsKit.poly_Fun — Method

poly_Fun(knots, ind; lb=-1.0, ub=1.0)

Lagrange interpolation polynomial ind for the set of knots

Arguments

knots::Vector{T}: A vector of knots

ind::Int: Knot index for polynomial equal to 1.

lb::Float64: Domain lower bound. Default is -1.0.

ub::Float64: Domain upper bound. Default is 1.0.

Returns

f::Fun: Lagrange interpolation polynomial as ApproxFun Fun object.

SparseGridsKit.reverse_kron_index — Method

reverse_kron_index(index::Int, dims::Vector{Int})

Given linear index index and a vector of dimensions dims, computes the multi-dimensional tensor index.

Arguments

index::Int: Linear index.

dims::Vector{Int}: Vector containing the storage dimension of the tensor dimensions.

Returns

Vector{Int}: Tensor index

SparseGridsKit.truncated_kron — Method

truncated_kron(vectors; tol=1e-10)

Compute the Kronecker product of a list of vectors with truncation.

Arguments

vectors::Vector{Vector}: Vector of vectors to compute the Kronecker product.

tol::Float64: Tolerance value for coefficienet truncation to promote sparsity. Elements with absolute values less than tol are set to zero. Default is 1e-10.

Returns

result: SparseVector representation of truncated Kronecker product of the input vectors.

tensor_dims::Vector{Int}: Dimensions of the resulting tensor.